requirements.txt
streamlit
google-generativeai
python-dotenv
python-docx
fpdf2
streamlit-lottie
